requirements.txt
python-dotenv
google-generativeai
pdf2image
six
pillow
pytesseract
gradio
requests